Optimize the writing of the image contents
Instead of writing each byte separately we now prepare a single buffer, and write that.
The buffer itself is filled from a pre-calculated look-up-table of the color values
(0...255) to the actual bytes that must be written. This includes padding, so when filling
the buffer we can exactly point to the index of a specific component.